+++ lib/rails_markup/cli/base.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "thor"
+
+module RailsMarkup
+  class Cli < Thor
+  end
+end