+++ test/dummy/config/environments/test.rb
@@ -6,4 +6,5 @@ Rails.application.configure do
   config.consider_all_requests_local = true
   config.action_dispatch.show_exceptions = :none
   config.action_controller.allow_forgery_protection = false
+  config.active_record.maintain_test_schema = false
 end