+++ test/dummy/db/schema.rb
@@ -12,7 +12,10 @@ ActiveRecord::Schema.define do
     t.text :selected_text
     t.text :metadata, default: "{}"
     t.text :thread, default: "[]"
+    t.string :client_uuid
 
     t.timestamps
   end
+
+  add_index :rails_markup_annotations, :client_uuid, unique: true
 end